% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bottom_temp_daily_epu.R
\name{make_bottom_temp_daily_epu}
\alias{make_bottom_temp_daily_epu}
\title{Make gridded daily files split by EPU for heatwave index}
\usage{
make_bottom_temp_daily_epu(
  input.file,
  output.file,
  shp.file,
  file.year,
  write.out = T
)
}
\arguments{
\item{input.file}{Either a character vector of full input file names for a list of spatrasters}

\item{output.file}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{file.year}{numeric. Year of the input file}

\item{write.out}{logical. If TRUE, the function will write out the gridded data to a netcdf file}
}
\value{
a dataframe (Time, Latitude, Longitude, var, value) or csv file of the gridded data
}
\description{
descriptions
}
