% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bottom_temp_daily_NEUS.R
\name{make_bottom_temp_daily_NEUS}
\alias{make_bottom_temp_daily_NEUS}
\title{Createst the bottom_temp_model_gridded (gridded seasonal means) inidcator for the State of the Ecosystem Report}
\usage{
make_bottom_temp_daily_NEUS(
  input.file,
  output.dir,
  output.prefix = NA,
  shp.file,
  write.out = T
)
}
\arguments{
\item{input.file}{Either a character vector of full input file names for a list of spatrasters}

\item{write.out}{logical. If TRUE, the function will write out the gridded data to a netcdf file}

\item{output.file}{character vector of full output file names corresponding to each input file}

\item{file.year}{numeric. Year of the input file}
}
\value{
a dataframe (Time, Latitude, Longitude, var, value) or csv file of the gridded data
}
\description{
descriptions
}
