% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bottom_temp_model_anom.R
\name{make_bottom_temp_model_anom}
\alias{make_bottom_temp_model_anom}
\title{Createst the bottom_temp_model_gridded (gridded seasonal means) inidcator for the State of the Ecosystem Report}
\usage{
make_bottom_temp_model_anom(
  input.file,
  output.file = NA,
  file.year,
  shp.file,
  climatology.file,
  write.out = F
)
}
\arguments{
\item{input.file}{Either a character vector of full input file names for a list of spatrasters}

\item{output.file}{character vector of full output file names corresponding to each input file}

\item{file.year}{numeric. Year of the input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{climatology.file}{string. File name for climatology file created by make_soe_climatology()}

\item{write.out}{logical. If TRUE, the function will write out the gridded data to a netcdf file}
}
\value{
a dataframe (Time, Latitude, Longitude, var, value) or csv file of the gridded data
}
\description{
descriptions
}
