% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bottom_temp_vast.R
\name{make_bottom_temp_temp_vast}
\alias{make_bottom_temp_temp_vast}
\title{Formats data for soe vast models}
\usage{
make_bottom_temp_temp_vast(
  input.index,
  output.dir,
  output.prefix,
  shp.file,
  redo = F
)
}
\arguments{
\item{input.index}{Dataframe of series|dir.name|start.year|end.year}

\item{output.dir}{Full output file directory}

\item{output.prefix}{string. Prefix for the output files}

\item{shp.file}{Full input file directory for the cold pool area shape file}

\item{origin}{string. Origin date for the time variable in the netCDF files (default '1970-01-01')}

\item{time.unit}{string. Time unit for the time variable in the netCDF files (default 'days')}
}
\value{
Annual RDS file with the processed data for GLORYS and ROMS (mod_bt longitude latitude time date year month day)
}
\description{
descriptions
}
