% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cold_pool_data.R
\name{make_cold_pool_data}
\alias{make_cold_pool_data}
\title{Formats data for cold pool analysis}
\usage{
make_cold_pool_data(input.files, output.dir, output.prefix, cp.shp.file)
}
\arguments{
\item{input.files}{Full input file directory for bottom temp data}

\item{output.dir}{Full output file directory}

\item{output.prefix}{string. Prefix for the output files}

\item{cp.shp.file}{Full input file directory for the cold pool area shape file}
}
\value{
RDS file with the processed data for GLORYS and ROMS
}
\description{
descriptions
}
