% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cold_pool_extent.R
\name{make_cold_pool_extent}
\alias{make_cold_pool_extent}
\title{Creates the maximum extent of the cold pool over the input data}
\usage{
make_cold_pool_extent(input.data, output.dir, shp.file, grid.file)
}
\arguments{
\item{input.data}{data.frame formatted by make_cold_pool_data}

\item{output.dir}{Full output file directory}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{grid.file}{Full input file directory for the cold pool grid file}
}
\value{
RDS file with the annual and maximum cold pool extent
}
\description{
descriptions
}
