% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cold_pool_extent_shp.R
\name{make_cold_pool_extent_shp}
\alias{make_cold_pool_extent_shp}
\title{Creates the cold pool extent shape file and metrics}
\usage{
make_cold_pool_extent_shp(input.files, output.dir)
}
\arguments{
\item{input.files}{Full input file directory for bottom temp data returned from make_cold_pool_data}

\item{output.dir}{Full output file directory}

\item{cp.extent.file}{Full input file directory for cold pool extent data returned from make_cold_pool_extent}
}
\value{
csv with cold pool indices
}
\description{
descriptions
}
