% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cold_pool_indices.R
\name{make_cold_pool_indices}
\alias{make_cold_pool_indices}
\title{Creates the cold pool index, persistence, and extent}
\usage{
make_cold_pool_indices(input.data, cp.extent.file, output.dir)
}
\arguments{
\item{input.data}{Full input file directory for bottom temp data returned from make_cold_pool_data}

\item{cp.extent.file}{Full input file directory for cold pool extent data returned from make_cold_pool_extent}

\item{output.dir}{Full output file directory}
}
\value{
csv with cold pool indices
}
\description{
descriptions
}
