% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cold_pool_shp.R
\name{make_cold_pool_shp}
\alias{make_cold_pool_shp}
\title{Creates the shape file for the maximum cold pool extent}
\usage{
make_cold_pool_shp(bathy.file, shp.file, output.file)
}
\arguments{
\item{bathy.file}{Full input file directory for bathymetry data}

\item{shp.file}{Full input file directory for the shape file to crop}

\item{output.file}{Full output file directory for the cold pool area shape file}
}
\value{
.shp with cold pool detection area
}
\description{
descriptions
}
