% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gridded_climatology.R
\name{make_gridded_climatology}
\alias{make_gridded_climatology}
\title{Creates a climatology}
\usage{
make_gridded_climatology(
  input.dir,
  input.prefix,
  output.file = NA,
  shp.file,
  ref.year.start,
  ref.year.end,
  agg.time,
  var.name,
  write.out = F
)
}
\arguments{
\item{input.dir}{Either a character vector of full input file names for a list of spatrasters}

\item{input.prefix}{string. Prefix for the input file names}

\item{output.file}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{ref.year.start}{numeric. Year that climatological period starts}

\item{ref.year.end}{numeric. Year that climatological period ends}

\item{agg.time}{string. Aggregation level. Options are 'season' or 'annual'}

\item{var.name}{string. Name of the variable in the input file}

\item{write.out}{logical. If TRUE, the function will write out the gridded data to a netcdf file}
}
\value{
a dataframe (Time, Latitude, Longitude, var, value) or csv file of the gridded data
}
\description{
descriptions
}
