% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_gridded.R
\name{make_monthly_gridded}
\alias{make_monthly_gridded}
\title{Creates a monthly gridded summary of gridded data inputs}
\usage{
make_monthly_gridded(
  input.dir,
  input.prefix,
  output.dir,
  output.prefix,
  var.name,
  statistics
)
}
\arguments{
\item{input.dir}{String. The full input directory for daily input files}

\item{input.prefix}{string. The prefix of the input files}

\item{output.dir}{string. The full output directory for the gridded data}

\item{output.prefix}{string. The prefix of the output files}

\item{var.name}{string. The variable name for the gridded data}

\item{statistics}{string. The statistic to be used for the gridded data}
}
\value{
a netCDF file with the monthly gridded data
}
\description{
descriptions
}
