% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_thermal_habitat_area.R
\name{make_thermal_habitat_area}
\alias{make_thermal_habitat_area}
\title{Creates the thermal_habitat_gridded indicator for the State of the Ecosystem Report}
\usage{
make_thermal_habitat_area(
  input.file,
  output.file.area = NA,
  output.file.gridded = NA,
  shp.file,
  file.year,
  t.max.seq,
  write.area = F,
  write.gridded = F
)
}
\arguments{
\item{input.file}{Character. Either a single path to a NetCDF file with multiple time layers,
or a character vector of file paths (e.g. daily files) to be combined.}

\item{output.file.area}{character with full file name for thermal_habitat_area output csv}

\item{output.file.gridded}{character with full file name for thermal_habitat_gridded output csv}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{file.year}{numeric. Year of the input file}

\item{t.max.seq}{numeric. A vector of temperature thresholds to calculate the number of days above}

\item{write.area}{logical. If TRUE, the function will write out the area csv file}

\item{write.gridded}{logical. If TRUE, the function will write out the gridded csv file}
}
\value{
a dataframe (Time, Latitude, Longitude, var, value) or csv file of the gridded data
}
\description{
descriptions
}
