% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_thermal_habitat_gridded.R
\name{make_thermal_habitat_gridded}
\alias{make_thermal_habitat_gridded}
\title{Createst the thermal_habitat_gridded inidcator for the State of the Ecosystem Report}
\usage{
make_thermal_habitat_gridded(
  input.file,
  output.file,
  shp.file,
  file.year,
  t.max.seq,
  write.out = F
)
}
\arguments{
\item{input.file}{Either a character vector of full input file names for a list of spatrasters}

\item{output.file}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{file.year}{numeric. Year of the input file}

\item{t.max.seq}{numeric. A vector of temperature thresholds to calculate the number of days above}

\item{write.out}{logical. If TRUE, the function will write out the gridded data to a netcdf file}
}
\value{
a dataframe (Time, Latitude, Longitude, var, value) or csv file of the gridded data
}
\description{
descriptions
}
